nova simd: call setvec_simd with float argument, not with double

git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/trunk@9484 a380766d-ff14-0410-b294-a243070f3f08

+++ Source/plugins/LFUGens.cpp
@@ -1629,7 +1629,7 @@ void XLine_next_nova(XLine *unit, int inNumSamples)
 
 	if (counter == 0)
 	{
-		nova::setvec_simd(OUT(0), level, inNumSamples);
+		nova::setvec_simd(OUT(0), (float)level, inNumSamples);
 		return;
 	}
 	if (counter >= inNumSamples)
